pyproject.toml
[project]
name = "nephele"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.4",
    "django-ninja>=1.3.0",
    "django-ninja-extra>=0.22.2",
    "psycopg[binary]>=3.2.3",
    "celery>=5.4.0",
    "cryptography>=44.0.0",
    "daphne>=4.1.2",
    "kubernetes>=31.0.0",
    "loguru>=0.7.3",
    "django-lifecycle>=1.2.4",
    "django-debug-toolbar>=5.0.1",
]

[tool.uv]
dev-dependencies = [
    "celery-types>=0.22.0",
    "django-types>=0.20.0",
    "django-stubs[compatible-mypy]>=5.1.1",
    "django-fake-model>=0.1.4",
    "kubernetes-stubs-elephant-fork>=31.0.0",
    "djlint>=1.36.4",
]
index-url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"

